memcached_dos/requirements.txt
click
python-memcached
scapy